:put_litter_in_its_place: Refactor duplicated update code
