+++ config/config.js
@@ -2,7 +2,8 @@ module.exports = {
   es: {
     host: process.env.ES_HOST || 'es',
     port: process.env.ES_PORT || '9200',
-    index: process.env.ES_INDEX || 'pharmacies'
+    index: process.env.ES_INDEX || 'pharmacies',
+    type: 'pharmacy',
   },
   timezone: 'Europe/London'
 };
