+++ test/integration/es/core-opening-times.js
@@ -1,7 +1,7 @@
 const chai = require('chai');
 const moment = require('moment');
 const esClient = require('../../../app/lib/esClient');
-const utils = require('./../testUtils');
+const utils = require('../../lib/testUtils');
 
 const expect = chai.expect;
