+++ test/integration/securityHeaders.js
@@ -1,7 +1,7 @@
 const chai = require('chai');
 const chaiHttp = require('chai-http');
 const app = require('../../app');
-const utils = require('./testUtils');
+const utils = require('../lib/testUtils');
 
 const expect = chai.expect;
