+++ test/unit/lib/config.js
@@ -6,6 +6,7 @@ const expect = chai.expect;
 describe('config', () => {
   it('should return pharmacies as the default collection', () => {
     expect(config.es.index).to.be.equal('pharmacies');
+    expect(config.es.type).to.be.equal('pharmacy');
     expect(config.es.host).to.be.equal('es');
     expect(config.es.port).to.be.equal('9200');
   });
